Semaforical control for Vissim using python
A code for semaforical control in traffic simulation. The software for traffic simulation used was Vissim, and the semaforical groups were controlled by COM interface in python. In this script, the semaforical states are changed every 30 seconds. For more information about Vissim and COM interface please refer to the following links: 
https://www.ptvgroup.com/en/solutions/products/ptv-vissim
https://www.youtube.com/watch?v=mhFI-DQZ5og